@@ -1,32 +1,46 @@
 source 'http://rubygems.org'
-ruby "1.9.3"
-gem 'rails', '3.2.13'
+ruby "2.1.5"
+gem 'rails', '4.2.0'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
-group :development do
-
-end
 
 # Gems used only for assets and not required
 # in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails',   '~> 5.0'
+  gem 'uglifier', '>= 1.3.0'
+  gem 'coffee-rails', '~> 4.1.0'
   gem "mini_magick"
-  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
-  # gem 'therubyracer', :platforms => :ruby
+  gem 'turbolinks'
+  gem 'therubyracer', :platforms => :ruby
   gem "bootstrap-sass"
-  gem 'uglifier', '>= 1.0.3'
+  gem 'jbuilder', '~> 2.0'
   gem "carrierwave"
   gem "nested_form"
    gem "jquery-fileupload-rails"
-end
-  gem 'pg', '0.15.1'
+  gem 'pg'
 gem 'simple_form'
 gem 'jquery-rails'
 gem "ransack"
+gem 'kristin'
+gem 'nokogiri'
+gem 'listen', '~> 2.7'
+gem 'wdm', '>= 0.1.0' if Gem.win_platform?
+gem 'sdoc', '~> 0.4.0', group: :doc
 
+gem 'tzinfo-data', platforms: [:x64_mingw, :mingw, :mswin]
+#gem 'spring',        group: :development
+
+group :development, :test do
+  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
+  gem 'byebug'
+
+  # Access an IRB console on exception pages or by using <%= console %> in views
+  gem 'web-console', '~> 2.0'
+
+  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'spring'
+end
 
 group :production do
   gem 'rails_12factor'
