@@ -1,4 +1,4 @@
-HousingDB::Application.routes.draw do
+Rails.application.routes.draw do
       
 
     post 'houses/edit_multiple'
@@ -7,25 +7,26 @@ HousingDB::Application.routes.draw do
     get "houses/viewall"
     get "houses/searchresults"
     get "houses/search"
-    
+    get "houses/pdfconverter"
   resources :housefiles  
   resources :houses do
     resources :listings
     resources :sales
     resources :housefiles
-    get'statusupdate', :on => :member
-    get'addattach', :on => :member
-    get'editattach', :on => :member
+    get'statusupdate', on: :member
+    get'addattach', on: :member
+    get'editattach', on: :member
+    post :pdfupload, :on => :collection
   end
 root to: 'houses#index'
   resources :agents
    resources :sales do 
          resources :agents
-   get'agentadd', :on => :member
+   get'agentadd', on: :member
     end
    resources :listings do 
          resources :agents
-   get'agentadd', :on => :member
+   get'agentadd', on: :member
     end
 
   # The priority is based upon order of creation:
@@ -77,7 +78,6 @@ root to: 'houses#index'
 
   # You can have the root of your site routed with "root"
   # just remember to delete public/index.html.
-  root :to => 'houses#index'
 
   # See how all your routes lay out with "rake routes"
 
