@@ -12,15 +12,16 @@ before_filter :get_house
  end
  
  def create
- @sale = @house.sales.new(params[:sale])
+   ad_params = params.require(:sale).permit!
+ @sale = @house.sales.new(ad_params)
 
   if @sale.save
 
   
-  if @sale.saledate >= @house.sales.find(:first, :order => "saledate DESC").saledate
-    if  @house.listings.find(:first, :order => "listingdate DESC") == nil
+  if @sale.saledate >= @house.sales.order('saledate DESC').first.saledate
+    if  @house.listings.order('listingdate DESC').first == nil
       @house.update_attributes(:currentprice => @sale.price, :status => 'Sold')
-    elsif @sale.saledate >= @house.listings.find(:first, :order => "listingdate DESC").listingdate
+    elsif @sale.saledate >= @house.listings.order('listingdate DESC').first.listingdate
       @house.update_attributes(:currentprice => @sale.price, :status => 'Sold')
     end    
   end
@@ -61,7 +62,8 @@ before_filter :get_house
      @sale.build_agent
 
     respond_to do |format|
-      if @sale.update_attributes(params[:sale])
+      ad_params = params.require(:sake).permit!
+      if @sale.update_attributes(ad_params)
         format.html { redirect_to house_sales_path(@sale.house_id), notice: 'Agent was successfully updated.' }
         format.json { head :no_content }
       else
