@@ -1,4 +1,4 @@
-HousingDB::Application.configure do
+Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
   # In the development environment your application's code is reloaded on
@@ -7,8 +7,8 @@ HousingDB::Application.configure do
   config.cache_classes = false
 
   # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
-
+  #config.whiny_nils = true
+  config.eager_load = false
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
@@ -20,20 +20,26 @@ HousingDB::Application.configure do
   config.active_support.deprecation = :log
 
   # Only use best-standards-support built into browsers
-  config.action_dispatch.best_standards_support = :builtin
-
+  #config.action_dispatch.best_standards_support = :builtin
+  config.active_record.migration_error = :page_load
   # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
+  #config.active_record.mass_assignment_sanitizer = :strict
 
   # Log the query plan for queries taking more than this (works
   # with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
+  #config.active_record.auto_explain_threshold_in_seconds = 0.5
 
   # Do not compress assets
-  config.assets.compress = false
+  #config.assets.compress = false
 
   # Expands the lines which load the assets
   config.assets.debug = true
   
-  config.serve_static_assets = true
+  config.assets.digest = true
+  
+  config.assets.raise_runtime_errors = true
+  
+  config.action_view.raise_on_missing_translations = true
+  
+  config.serve_static_files = true
 end
